Add authentication to mutations

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -1,4 +1,7 @@
 import bcrypt from 'bcryptjs';
+import jwt from 'jsonwebtoken';
+
+import getUserId from '../utils/getUserId';
 
 const Mutation = {
     async createUser(parent, args, {
@@ -9,40 +12,88 @@ const Mutation = {
 
         const password = await bcrypt.hash(args.data.password, 10);
 
-        return prisma.mutation.createUser({
+        const user = prisma.mutation.createUser({
             data: {
                 ...args.data,
                 password
             }
-        }, info);
+        });
+        return {
+            user,
+            token: jwt.sign({
+                userId: user.id
+            }, 'thisisasecret')
+        }
     },
-    updateUser(parent, args, {
+    async loginUser(parent, {
+        email,
+        password
+    }, {
         prisma
     }, info) {
+        const user = await prisma.query.user({
+            where: {
+                email
+            }
+        });
+        if (!user)
+            throw new Error('Invalid email or password');
+
+        const isMatch = await bcrypt.compare(password, user.password);
+        if (!isMatch)
+            throw new Error('Invalid email or password');
+
+        return {
+            user,
+            token: jwt.sign({
+                userId: user.id
+            }, 'thisisasecret')
+        }
+    },
+    updateUser(parent, args, {
+        prisma,
+        request
+    }, info) {
+        const userId = getUserId(request);
         return prisma.mutation.updateUser({
             where: {
-                id: args.id
+                id: userId
             },
             data: args.data
         }, info);
     },
     createPost(parent, args, {
-        prisma
+        prisma,
+        request
     }, info) {
+        const userId = getUserId(request);
         return prisma.mutation.createPost({
             data: {
                 ...args.data,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 }
             }
         }, info);
     },
-    updatePost(parent, args, {
-        prisma
+    async updatePost(parent, args, {
+        prisma,
+        request
     }, info) {
+        const userId = getUserId(request);
+
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        });
+
+        if (!postExists)
+            throw new Error('Could not update post')
+
         return prisma.mutation.updatePost({
             where: {
                 id: args.id
@@ -51,14 +102,17 @@ const Mutation = {
         }, info);
     },
     createComment(parent, args, {
-        prisma
+        prisma,
+        request
     }, info) {
+        const userId = getUserId(request);
+
         return prisma.mutation.createComment({
             data: {
                 text: args.data.text,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 },
                 post: {
@@ -69,9 +123,22 @@ const Mutation = {
             }
         }, info);
     },
-    updateComment(parent, args, {
-        prisma
+    async updateComment(parent, args, {
+        prisma,
+        request
     }, info) {
+        const userId = getUserId(request);
+
+        const commmentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        });
+
+        if (!commmentExists)
+            throw new Error('Could not update comment');
+
         return prisma.mutation.updateComment({
             where: {
                 id: args.id
@@ -80,26 +147,53 @@ const Mutation = {
         }, info);
     },
     deleteUser(parent, args, {
-        prisma
+        prisma,
+        request
     }, info) {
+        const userId = getUserId(request);
         return prisma.mutation.deleteUser({
             where: {
-                id: args.id
+                id: userId
             }
         }, info);
     },
-    deletePost(parent, args, {
-        prisma
+    async deletePost(parent, args, {
+        prisma,
+        request
     }, info) {
+        const userId = getUserId(request);
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        });
+
+        if (!postExists)
+            throw new Error('Unable to delete post');
+
         return prisma.mutation.deletePost({
             where: {
                 id: args.id
             }
         }, info);
     },
-    deleteComment(parent, args, {
-        prisma
+    async deleteComment(parent, args, {
+        prisma,
+        request
     }, info) {
+        const userId = getUserId(request);
+
+        const commmentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        });
+
+        if (!commmentExists)
+            throw new Error('Could not delete comment');
+
         return prisma.mutation.deleteComment({
             where: {
                 id: args.id
